Renamed components and fixed heading
